Increase gateway retry interval, retry max, and request timeouts (#233)
* Increase gateway retry interval, retry max, and request timeouts

Serverless instances can take more than 5 minutes to start up so when jupyter server attempts to connect to the kernel, the default gateway interval and max can cause it to timeout and drop messages.

Batch creation is done via a POST request and can take up to 10 minutes to be created (POST response is returned). This change increases that request timeout to be 10 minutes.

All of these values are treated as minimums and will be set if and only if the config setting is not set or the config setting is less than the min value.

* Update the set action to be more pythonic